+++ src/base/core/src/num/dom/mod.rs
@@ -1,31 +1,44 @@
 // devela_base_core::num::dom
 //
 #![doc = crate::_DOC_NUM_DOM!()]
+#![doc = crate::_doc!(modules: crate::num; dom: int, real)]
+#![doc = crate::_doc!(flat:"num")]
+//!
+#![doc = crate::_doc!(flat:"num")]
 //
 
+mod _internals; // impl_ops!, upcasted_op! TODO:RENAME
+
 // mod complex;
-mod int; // Divisor, GcdReturn, Int, [i|u]size_[down|up]
 // mod fract;
 // mod laws;
 // mod ops;
-mod real; // Real-valued numeric domains and representations.
 mod traits; // NumConst
 
-crate::structural_mods! { // _mods, _workspace_internals, _hidden
+pub mod int; // Divisor, GcdReturn, Int, [i|u]size_[down|up]
+pub mod real; // Real-valued numeric domains and representations.
+
+crate::structural_mods! { // _mods, _pub_mods, _workspace_internals, _hidden
     _mods {
         #[doc(inline)]
         pub use super::{
             // complex::_all::*,
-            int::_all::*,
             // fract::_all::*,
             // laws::_all::*,
             // ops::_all::*,
-            real::_all::*,
             traits::_all::*,
+        };
+    }
+    _pub_mods {
+        #[doc(inline)]
+        pub use super::{
+            int::_all::*,
+            real::_all::*,
         };
     }
     _workspace_internals {
         pub use super::{
+            _internals::*,
             real::_workspace_internals::*,
             int::_workspace_internals::*,
         };