+++ src/num/dom/frac/mod.rs
@@ -6,7 +6,6 @@
 #[cfg(feature = "int")]
 mod wrapper;
 
-// WIPZONE
 // mod r#trait;
 
 crate::structural_mods! { // _mods
@@ -14,7 +13,6 @@ crate::structural_mods! { // _mods
         #[cfg(feature = "int")]
         pub use super::wrapper::*;
 
-        // WIPZONE
         // pub use super::r#trait::*;
     }
 }