+++ src/num/dom/frac/wrapper/mod.rs
@@ -1,4 +1,4 @@
-// devela::num::frac::wrapper
+// devela::num::dom::frac::wrapper
 //
 //! Fraction-related wrapper struct.
 //
@@ -13,7 +13,7 @@ mod impl_frac;
 
 #[doc = crate::_tags!(num namespace)]
 /// Provides comprehensive fractional operations on `T`, most of them *const*.
-#[doc = crate::_doc_location!("num")]
+#[doc = crate::_doc_location!("num/dom")]
 ///
 /// It's implemented for:
 /// - arrays: `[i8; 2]`… `[u128; 2]`; `[Int<i8>; 2]`… `[Int<u128>; 2]`.