+++ src/num/dom/mod.rs
@@ -1,27 +1,35 @@
 // devela::num::dom
 //
 #![doc = crate::_DOC_NUM_DOM!()]
+#![doc = crate::_doc!(modules: crate::num; dom: int, real)]
+#![doc = crate::_doc!(flat:"num")]
 //
 
 // mod complex;
-mod int; // NumInt[Ref], prime_number_theorem, (Divisor, GcdReturn, Int[Alloc], [i|u]size_[down|up])
-// mod fract;
+mod frac;
 // mod laws;
 // mod ops;
-mod real; // Real-valued numeric domains and representations.
 mod traits; // Num, NumRef, (NumConst)
 
-crate::structural_mods! { // _mods, _workspace_internals, _hidden
+pub mod int; // NumInt[Ref], prime_number_theorem, (Divisor, GcdReturn, Int[Alloc], [i|u]size_[down|up])
+pub mod real; // Real-valued numeric domains and representations.
+
+crate::structural_mods! { // _mods, _pub_mods
     _mods {
         #[doc(inline)]
         pub use super::{
             // complex::_all::*,
-            int::_all::*,
-            // fract::_all::*,
+            frac::_all::*,
             // laws::_all::*,
             // ops::_all::*,
-            real::_all::*,
             traits::_all::*,
+        };
+    }
+    _pub_mods {
+        #[doc(inline)]
+        pub use super::{
+            int::_all::*,
+            real::_all::*,
         };
     }
 }