+++ src/num/error/mod.rs
@@ -1,0 +1,32 @@
+// devela::num::error
+//
+//! Numeric-related error types.
+//!
+#![doc = crate::_doc!(flat:"num")]
+//
+
+mod definitions;
+
+crate::structural_mods! { // _mods, _reexports
+    _mods {
+        #[doc(inline)]
+        pub use super::{
+            definitions::*,
+        };
+    }
+    _reexports {
+        pub use devela_base_core::num::error::{
+            // individual errors:
+            IncompatibleBounds,
+            NoInverse,
+            MismatchedSizes,
+            NonNegativeRequired,
+            PositiveRequired,
+            NonZeroRequired,
+            Overflow,
+            // composite errors:
+            IntError, IntResult,
+            NicheValueError,
+        };
+    }
+}