+++ src/phys/unit/_helpers.rs
@@ -1,4 +1,4 @@
-// devela::num::unit::helpers
+// devela::phys::unit::_helpers
 
 // Implements TryFrom for `$to_prim` from the `$unit` type `$from_prim` value
 macro_rules! impl_try_from {